Move deferred transaction into deferred_transaction.hpp
